+++ packages/fabric/editor-confluence-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-cq-transformer
 
+## 2.7.6
+
 ## 2.7.5
 - [patch] bump editor-common to 6.1.2 [bb7802e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/bb7802e)