+++ packages/fabric/media-avatar-picker/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/media-avatar-picker
 
+## 5.0.1
+
 ## 5.0.0
 
 ## 4.6.0